Refactor and update invoice and customer models
- Removed unnecessary using directives in `Customer.cs` and `Store.cs`.
- Added a private constructor to `SmsMessage` class.
- Changed `DueAmount` in `Invoice` to a private set property.
- Updated `TotalPrice` in `InvoiceItem` to have a private setter.
- Renamed `CalculateTotalPrice` to `UpdateTotalPrice` and modified its implementation.
- Updated methods in `InvoiceItem` to call `UpdateTotalPrice` after setting values.
- Changed `ProductId` type from `int` to `long` in `ProductVariant`.
- Updated `PhoneNumber` property to `Phone` in `Store` and adjusted related properties.
- Modified `CustomerEntityTypeConfiguration` to reflect changes in the `Phone` property and added a uniqueness index.
- Changed column type for `ParametersJson` in configuration files from `jsonb` to `NVARCHAR(MAX)`.
- Added a new folder for `Migrations` in the project structure.
- Removed commented-out code for `AddCustomerViewModel`.
- Added migration files for the initial database schema.
- Updated `ApplicationDbContextModelSnapshot` to reflect the current schema state.

+++ StoreManagement.Domain/Aggregates/Products/ProductVariant.cs
@@ -8,13 +8,13 @@ public class ProductVariant : BaseEntity
     public string SKU { get; private set; }
     public string? Color { get; private set; }
     public string? Size { get; private set; }
-    public int ProductId { get; private set; }
+    public long ProductId { get; private set; }
     public Product Product { get; private set; }
 
     private ProductVariant() { }
 
     // Constructor for creating a new ProductVariant
-    public ProductVariant(string sku, string? color, string? size, int productId)
+    public ProductVariant(string sku, string? color, string? size, long productId)
     {
         SKU = sku ?? throw new ArgumentNullException(nameof(sku));
         Color = color ?? throw new ArgumentNullException(nameof(color));
